Enable github.security to be run standalone
This moves the entrypoint to the self-contained security.py module, which is
then called directly from the dokku cron job.
